requirements.txt
pandas~=1.1.3
scrapy~=2.4.1

sqlalchemy~=1.3.20
python-dotenv~=0.15.0
typer~=0.3.2
tqdm~=4.51.0